requirements.txt
amqp==5.1.1
asgiref==3.7.2
billiard==4.1.0
celery==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
Django==4.2.3
flake8==6.1.0
gunicorn==21.2.0
idna==3.4
kombu==5.3.1
marshmallow==3.20.1
marshmallow-dataclass==8.5.14
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.1
prompt-toolkit==3.0.39
psycopg2-binary==2.9.6
pycodestyle==2.11.0
pyflakes==3.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
redis==4.6.0
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
vine==5.0.0
wcwidth==0.2.6
whitenoise==6.5.0
